compiler: the compile entry points move out of the CLI layer
compile/compileMany/resetCompilationCache (the default-session
module-level entry points) move from lib/cli/commands.ts to a new
lib/compiler/defaultSession.ts, next to the BuildSession they delegate
to. This removes the last upward import in the library layer -
lib/eval/run/seed.ts no longer reaches into lib/cli. The lazy-session
construction (and its pack tree-shaking rationale) moves verbatim.
All importers repointed; no re-export shims.
