+++ packages/agency-lang/lib/cli/coverage.ts
@@ -3,7 +3,7 @@ import * as path from "path";
 import { pathToFileURL } from "url";
 import picomatch from "picomatch";
 import { AgencyConfig } from "../config.js";
-import { compile } from "./commands.js";
+import { compile } from "@/compiler/defaultSession.js";
 import { RunStrategy } from "../importStrategy.js";
 import { ttyColor } from "../utils/termcolors.js";
 import renderCoverageHtml from "../templates/cli/coverageReport.js";
