+++ packages/agency-lang/lib/cli/serve.ts
@@ -2,7 +2,8 @@ import fs from "fs";
 import path from "path";
 import process from "process";
 import { fileURLToPath, pathToFileURL } from "url";
-import { compile, loadConfig } from "./commands.js";
+import { loadConfig } from "./commands.js";
+import { compile } from "@/compiler/defaultSession.js";
 import { formatErrors } from "../typeChecker/index.js";
 import { collectServeMetadata } from "../serve/metadata.js";
 import { discoverExports } from "../serve/discovery.js";
